# my-auth-api/docs/swagger.yaml
basePath: /
definitions:
  main.User:
    properties:
      id:
        type: integer
      password:
        type: string
      username:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: API untuk otentikasi user dengan JWT di Golang menggunakan Gin Framework.
  title: MyAuth API
  version: "1.0"
paths:
  /api/profile:
    get:
      description: Mengambil profil user berdasarkan token JWT
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get User Profile
      tags:
      - User
  /login:
    post:
      consumes:
      - application/json
      description: Login user dan mendapatkan token JWT
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login User
      tags:
      - Auth
  /logout:
    post:
      description: Logout user dengan menghapus token dari frontend
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Logout User
      tags:
      - Auth
  /register:
    post:
      consumes:
      - application/json
      description: Mendaftarkan user baru ke database
      parameters:
      - description: User Data
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/main.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Register User
      tags:
      - Auth
swagger: "2.0"